Another try to get the lint warnings fixed. The last attempt made things windows specific

+++ wispers-connect/src/juice.rs
@@ -75,7 +75,8 @@ impl State {
             x if x == ffi::juice_state_JUICE_STATE_CONNECTED => State::Connected,
             x if x == ffi::juice_state_JUICE_STATE_COMPLETED => State::Completed,
             x if x == ffi::juice_state_JUICE_STATE_FAILED => State::Failed,
-            other => State::Unknown(other),
+            #[allow(clippy::unnecessary_cast)]
+            other => State::Unknown(other as i32),
         }
     }
